+++ inst/include/epinow2.hpp
@@ -5,5 +5,6 @@
 // Must be included before the model code.
 #include "epinow2/convolve_with_rev_pmf.hpp"
 #include "epinow2/renewal_infections.hpp"
+#include "epinow2/R_to_r.hpp"
 
 #endif